adding CategoryRepository

+++ src/main/java/pl/sdacademy/entities/Category.java
@@ -8,20 +8,20 @@ import javax.persistence.*;
 public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String name;
     private String producent;
 
     public Category() {
     }
 
-    public Category(Integer id, String name, String producent) {
+    public Category(Long id, String name, String producent) {
         this.id = id;
         this.name = name;
         this.producent = producent;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
@@ -33,7 +33,7 @@ public class Category {
         return producent;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
